Fix #144982: Crash when scrubbing retimed strips
Caused by casting float value to int in source cache eviction logic. The cache
key was changed from int to float for timeline frame, but this was not changed
in `source_image_cache_evict()`. Compiler did not catch this unfortunately.

Pull Request: https://projects.blender.org/blender/blender/pulls/145019

+++ source/blender/sequencer/intern/cache/source_image_cache.cc
@@ -280,7 +280,7 @@ bool source_image_cache_evict(Scene *scene)
    * biasing the ones that are behind the current frame. */
   const int cur_frame = scene->r.cfra;
   SourceImageCache::StripEntry *best_strip = nullptr;
-  std::pair<int, int> best_key = {};
+  std::pair<float, int> best_key = {};
   int best_score = 0;
   for (const auto &strip : cache->map_.items()) {
     for (const auto &entry : strip.value.frames.items()) {
